Split up tests into integration and unit, run integration in parallel
