+++ .github/workflows/test-matrix-agp-gradle.yaml
@@ -1,4 +1,4 @@
-name: Test Publishing UUID Dry Run
+name: Test Matrix
 
 on:
   push:
@@ -40,7 +40,7 @@ jobs:
             gradle: "8.1"
             java: "17"
 
-    name: Publish Dry Run - AGP ${{ matrix.agp }} - Gradle ${{ matrix.gradle }}
+    name: Test Matrix - AGP ${{ matrix.agp }} - Gradle ${{ matrix.gradle }}
     env:
       VERSION_AGP: ${{ matrix.agp }}
       SENTRY_AUTH_TOKEN: ${{ secrets.SENTRY_AUTH_TOKEN }}
@@ -48,6 +48,12 @@ jobs:
     steps:
       - name: Checkout Repo
         uses: actions/checkout@v3
+
+      - name: Setup Gradle
+        uses: gradle/gradle-build-action@7e48093f71aa12588241894ff3695e83c4b5e4b0 # pin@v2.4.2
+        with:
+          cache-read-only: ${{ github.ref != 'refs/heads/main' }}
+          gradle-version: ${{ matrix.gradle }}
 
       - name: Setup Java Version
         uses: actions/setup-java@v3
@@ -57,18 +63,10 @@ jobs:
 
       - name: Run ./gradlew (Fixes Gradle < 7.4)
         continue-on-error: true
-        uses: gradle/gradle-build-action@7e48093f71aa12588241894ff3695e83c4b5e4b0 # pin@v2.4.2
-        with:
-          cache-read-only: ${{ github.ref != 'refs/heads/main' }}
-          gradle-version: ${{ matrix.gradle }}
-          arguments: tasks
+        run: ./gradlew tasks
 
       - name: Build the Release variant
-        uses: gradle/gradle-build-action@7e48093f71aa12588241894ff3695e83c4b5e4b0 # pin@v2.4.2
-        with:
-          cache-read-only: ${{ github.ref != 'refs/heads/main' }}
-          gradle-version: ${{ matrix.gradle }}
-          arguments: assembleRelease
+        run: ./gradlew assembleRelease
 
       - name: Check sentry-debug-meta.properties inside APKs
         run: |
@@ -80,11 +78,7 @@ jobs:
           rm -r output
 
       - name: Build the Release Bundle variant
-        uses: gradle/gradle-build-action@7e48093f71aa12588241894ff3695e83c4b5e4b0 # pin@v2.4.2
-        with:
-          cache-read-only: ${{ github.ref != 'refs/heads/main' }}
-          gradle-version: ${{ matrix.gradle }}
-          arguments: bundleRelease
+        run: ./gradlew bundleRelease
 
       - name: Check sentry-debug-meta.properties inside App Bundle
         run: |
@@ -94,3 +88,13 @@ jobs:
           unzip ./examples/android-gradle-kts/build/outputs/bundle/release/android-gradle-kts-release.aab -d output
           cat output/base/assets/sentry-debug-meta.properties
           rm -r output
+
+      - name: Run Integration Tests
+        run: ./gradlew integrationTest
+
+      - name: Upload Test Results
+        uses: actions/upload-artifact@v3
+        with:
+          name: test-results-${{ matrix.os }}
+          path: plugin-build/build/reports/tests/
+        if: always()
