+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginAutoInstallTest.kt
@@ -1,17 +1,16 @@
 package io.sentry.android.gradle
 
+import io.sentry.BuildConfig
 import io.sentry.android.gradle.SentryPlugin.Companion.SENTRY_SDK_VERSION
+import org.gradle.util.GradleVersion
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
-@RunWith(Parameterized::class)
-class SentryPluginAutoInstallTest(
-    androidGradlePluginVersion: String,
-    gradleVersion: String
-) : BaseSentryPluginTest(androidGradlePluginVersion, gradleVersion) {
+class SentryPluginAutoInstallTest :
+    BaseSentryPluginTest(BuildConfig.AgpVersion, GradleVersion.current().version) {
 
     @Test
     fun `adds sentry-android dependency`() {
