+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginCheckAndroidSdkTest.kt
@@ -1,16 +1,14 @@
 package io.sentry.android.gradle
 
+import io.sentry.BuildConfig
+import org.gradle.util.GradleVersion
 import kotlin.test.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
-@Suppress("FunctionName")
-@RunWith(Parameterized::class)
-class SentryPluginCheckAndroidSdkTest(
-    androidGradlePluginVersion: String,
-    gradleVersion: String
-) : BaseSentryPluginTest(androidGradlePluginVersion, gradleVersion) {
+class SentryPluginCheckAndroidSdkTest :
+    BaseSentryPluginTest(BuildConfig.AgpVersion, GradleVersion.current().version) {
 
     @Test
     fun `when tracingInstrumentation is disabled does not check sentry-android sdk state`() {
