+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginSourceContextNonAndroidTest.kt
@@ -1,17 +1,16 @@
 package io.sentry.android.gradle
 
+import io.sentry.BuildConfig
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import org.gradle.testkit.runner.TaskOutcome.SKIPPED
+import org.gradle.util.GradleVersion
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
-@RunWith(Parameterized::class)
-class SentrySourceContextTest(
-    androidGradlePluginVersion: String,
-    gradleVersion: String
-) : BaseSentryPluginTest(androidGradlePluginVersion, gradleVersion) {
+class SentryPluginSourceContextNonAndroidTest :
+    BaseSentryNonAndroidPluginTest(GradleVersion.current().version) {
 
     @Test
     fun `skips bundle and upload tasks if no sources`() {
@@ -19,16 +18,8 @@ class SentrySourceContextTest(
             // language=Groovy
             """
             plugins {
-              id "com.android.application"
-              id "io.sentry.android.gradle"
-            }
-
-            android {
-              namespace 'com.example'
-
-              buildFeatures {
-                buildConfig false
-              }
+              id "org.jetbrains.kotlin.jvm"
+              id "io.sentry.jvm.gradle"
             }
 
             sentry {
@@ -37,15 +28,15 @@ class SentrySourceContextTest(
             """.trimIndent()
         )
         val result = runner
-            .appendArguments("app:assembleRelease")
+            .appendArguments("app:assemble")
             .build()
 
         assertEquals(
-            result.task(":app:sentryUploadSourceBundleRelease")?.outcome,
+            result.task(":app:sentryUploadSourceBundleJava")?.outcome,
             SKIPPED
         )
         assertEquals(
-            result.task(":app:sentryBundleSourcesRelease")?.outcome,
+            result.task(":app:sentryBundleSourcesJava")?.outcome,
             SKIPPED
         )
         assertTrue { "BUILD SUCCESSFUL" in result.output }
@@ -57,16 +48,8 @@ class SentrySourceContextTest(
             // language=Groovy
             """
             plugins {
-              id "com.android.application"
-              id "io.sentry.android.gradle"
-            }
-
-            android {
-              namespace 'com.example'
-
-              buildFeatures {
-                buildConfig false
-              }
+              id "org.jetbrains.kotlin.jvm"
+              id "io.sentry.jvm.gradle"
             }
 
             sentry {
@@ -83,14 +66,13 @@ class SentrySourceContextTest(
 
         sentryPropertiesFile.writeText("")
 
-        val ktContents = testProjectDir.withDummyComposeFile()
+        val ktContents = testProjectDir.withDummyKtFile()
         val javaContents = testProjectDir.withDummyJavaFile()
         val customContents = testProjectDir.withDummyCustomFile()
 
         val result = runner
-            .appendArguments("app:assembleRelease")
+            .appendArguments("app:assemble")
             .build()
-
         assertTrue { "\"--org\" \"sentry-sdks\"" in result.output }
         assertTrue { "\"--project\" \"sentry-android\"" in result.output }
         assertTrue { "BUILD SUCCESSFUL" in result.output }
@@ -98,17 +80,23 @@ class SentrySourceContextTest(
         verifySourceBundleContents(
             testProjectDir.root,
             "files/_/_/com/example/Example.jvm",
-            ktContents
+            ktContents,
+            variant = "java",
+            archivePath = "app/build/libs/app.jar"
         )
         verifySourceBundleContents(
             testProjectDir.root,
             "files/_/_/com/example/TestJava.jvm",
-            javaContents
+            javaContents,
+            variant = "java",
+            archivePath = "app/build/libs/app.jar"
         )
         verifySourceBundleContents(
             testProjectDir.root,
             "files/_/_/io/other/TestCustom.jvm",
-            customContents
+            customContents,
+            variant = "java",
+            archivePath = "app/build/libs/app.jar"
         )
     }
 }
