+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginTest.kt
@@ -1,6 +1,8 @@
 package io.sentry.android.gradle
 
+import io.sentry.BuildConfig
 import io.sentry.android.gradle.extensions.InstrumentationFeature
+import org.gradle.util.GradleVersion
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotEquals
@@ -10,12 +12,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
-@Suppress("FunctionName")
-@RunWith(Parameterized::class)
-class SentryPluginTest(
-    androidGradlePluginVersion: String,
-    gradleVersion: String
-) : BaseSentryPluginTest(androidGradlePluginVersion, gradleVersion) {
+class SentryPluginTest :
+    BaseSentryPluginTest(BuildConfig.AgpVersion, GradleVersion.current().version) {
 
     @Test
     fun `plugin does not configure tasks`() {
