+++ plugin-build/src/test/kotlin/io/sentry/android/gradle/SentryPluginVariantTest.kt
@@ -1,17 +1,15 @@
 package io.sentry.android.gradle
 
+import io.sentry.BuildConfig
+import org.gradle.util.GradleVersion
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 
-@Suppress("FunctionName")
-@RunWith(Parameterized::class)
-class SentryPluginVariantTest(
-    androidGradlePluginVersion: String,
-    gradleVersion: String
-) : BaseSentryPluginTest(androidGradlePluginVersion, gradleVersion) {
+class SentryPluginVariantTest :
+    BaseSentryPluginTest(BuildConfig.AgpVersion, GradleVersion.current().version) {
 
     override val additionalRootProjectConfig: String =
         // language=Groovy
